+++ events/messageReactionRemove/removeReaction.js
@@ -1,4 +1,4 @@
-const Reaction = require('../../classes/reactiontest');
+const Reaction = require('../../classes/reaction');
 
 module.exports = async (reaction, user) => {
     await Reaction.messageReactionHandler(reaction, user, false);
